refactor: Add CVTFFile::Init, vlTrue/False -> constexpr instead of macro
Rename one variant of ::Create to ::Init. This was done because it was
basically initting an empty image instead of creating it from RGBA8888
data like the other Create variants.

Allow mip count to be manually specified in calls to Init.
CVTFFile::Create did not allow this, instead took a bool asking if mips
were enabled, and forced 10. That's problematic if you want to limit the
number of mips or use a smaller pot or non-pot image.